Add PlanetsRendererProtocol
